+++ lib/source_server.dart
@@ -2,6 +2,7 @@ export 'src/exceptions/exceptions.dart';
 export 'src/query/models/query_player.dart';
 export 'src/query/models/server_info.dart';
 export 'src/query/models/server_os.dart';
+export 'src/query/models/server_rule.dart';
 export 'src/query/models/server_type.dart';
 export 'src/query/models/server_vac.dart';
 export 'src/query/models/server_visibility.dart';